--- a/addons/web/static/src/js/core/utils.js
+++ b/addons/web/static/src/js/core/utils.js
@@ -268,6 +268,25 @@ var utils = {
         return Math.max(min, Math.min(max, val));
     },
     /**
+     * Looks through the list and returns the first value that matches all
+     * of the key-value pairs listed in properties.
+     * If no match is found, or if list is empty, undefined will be returned.
+     *
+     * @param {Array} list
+     * @param {Object} props
+     * @returns {any|undefined} first element in list that matches all props
+     */
+    findWhere: function (list, props) {
+        if (!Array.isArray(list) || !props) {
+            return;
+        }
+        return list.filter((item) => item !== undefined).find((item) => {
+            return Object.keys(props).every((key) => {
+                return item[key] === props[key];
+            })
+        });
+    },
+    /**
      * @param {number} value
      * @param {integer} decimals
      * @returns {boolean}