https://github.com/odoo/enterprise/pull/18366
https://github.com/odoo/documentation/pull/978
Taskid: 2412172

**Description of the issue/feature this PR addresses:**
With 500 employees, the gantt view tries to load more than 20.000 work entries. 

The browser painfully dies while loading the DOM. (1-2 seconds to load the records, 1 minute to load the DOM)

We already did some optimisations with VSC, but this is cleary not enough, and as the amount of employees is increasing, it will become more and more difficult to display the work entries.

I see 2 more ways to optimise the loading of the view:
    1/ Add the possibility to disable the drag and drop (that add a lot of useless stuff on each box), which is not useful for the work entries.
   2/ Add the possibility to display a pager (opt-in too) ==> _not retained as this feature is not a simple task, it should have its own task_

**Current behavior before PR:**

Tested on my local machine with 2000 employees - 60000 work entries.
The browser loads the DOM within 125 seconds (in debug=assets mode).

**Desired behavior after PR is merged:**
Tested on my local machine with 2000 employees - 60000 work entries.
The browser loads the DOM within 40 seconds (in debug=assets mode). Still slow, but 3x better.
An option has also been added to disable the drag&drop feature on a gantt view, which cut off by 5 sec when activated: 35sec in total.
Improvements could also be made on the python model side. `search_read` takes 15 to 20 seconds with this work entries count.

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr